move Dependency children out
