+++ simple/src/main/java/com/lezhnin/yadi/simple/SimplePostConstruct.java
@@ -3,12 +3,11 @@ package com.lezhnin.yadi.simple;
 import static java.util.Arrays.stream;
 import static java.util.Objects.requireNonNull;
 import com.lezhnin.yadi.api.Dependency;
-import com.lezhnin.yadi.api.Dependency.InstanceMethodDependency;
-import com.lezhnin.yadi.api.Dependency.MethodDependency;
+import com.lezhnin.yadi.api.InstanceMethodDependency;
+import com.lezhnin.yadi.api.MethodDependency;
 import com.lezhnin.yadi.api.MethodNotFoundException;
 import com.lezhnin.yadi.api.ServiceLocator;
 import com.lezhnin.yadi.api.ServiceReference;
-import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
 import javax.annotation.Nonnull;
